requirements.txt
loguru~=0.6.0
streamlit~=1.16.0
pillow~=9.2.0